web/requirements.txt
asgiref==3.11.1
black==26.1.0
click==8.3.1
Django==5.2.11
django-debug-toolbar==6.2.0
django-environ==0.13.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
isort==8.0.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.2
psycopg2-binary==2.9.11
PyJWT==2.11.0
pytokens==0.4.1
sqlparse==0.5.5
